synthesis should be compilable on Mac now
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@2722 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Synthesis/synthesis/trunk/gridding/WDependentGridderBase.h
@@ -50,7 +50,7 @@ public:
    /// @details
    /// @param[in] wmax Maximum baseline (wavelengths)
    /// @param[in] nwplanes Number of w planes   
-   WDependentGridderBase(const double wmax, const double nwplanes); 
+   WDependentGridderBase(const double wmax, const int nwplanes); 
    
    /// @brief obtain the number of w-planes
    /// @details
